[chore] hardware popup helper 메소드 추출
- 최초 생성 후 내부 조작이 필요한 것도 아니라서 책임 분리
