+++ src/class/hw.ts
@@ -2,7 +2,7 @@
 
 import HardwareSocketMessageHandler from './hardware/hardwareSocketMessageHandler';
 import HardwareMonitor from './hardware/hardwareMonitor';
-import PopupHelper from './popup_helper';
+import createHardwarePopup from './hardware/functions/createHardwarePopup';
 
 enum HardwareModuleType {
     builtIn = 'builtin',
@@ -356,7 +356,9 @@ export default class Hardware implements IEntry.Hardware {
 
     private _initHardwareObject() {
         const { hardwareEnable } = Entry;
-        this._hwPopupCreate();
+        this.popupHelper = createHardwarePopup(() => {
+            this.downloadConnector();
+        });
         hardwareEnable && this._initSocket();
     }
 
@@ -813,72 +815,6 @@ export default class Hardware implements IEntry.Hardware {
                 document.getElementsByTagName('body')[0].removeChild(iFrame);
             }, 500);
         }
-    }
-
-    private _hwPopupCreate() {
-        const hw = this;
-        if (!this.popupHelper) {
-            if (window.popupHelper) {
-                this.popupHelper = window.popupHelper;
-            } else {
-                this.popupHelper = new PopupHelper(true);
-            }
-        }
-
-        this.popupHelper.addPopup('hwDownload', {
-            type: 'confirm',
-            title: Lang.Msgs.not_install_title,
-            setPopupLayout(popup: any) {
-                const content = Entry.Dom('div', {
-                    class: 'contentArea',
-                });
-                const text = Entry.Dom('div', {
-                    class: 'textArea',
-                    parent: content,
-                });
-                const text1 = Entry.Dom('div', {
-                    class: 'text1',
-                    parent: text,
-                });
-                const text2 = Entry.Dom('div', {
-                    class: 'text2',
-                    parent: text,
-                });
-                const text3 = Entry.Dom('div', {
-                    class: 'text3',
-                    parent: text,
-                });
-                const text4 = Entry.Dom('div', {
-                    class: 'text4',
-                    parent: text,
-                });
-                const cancel = Entry.Dom('div', {
-                    classes: ['popupCancelBtn', 'popupDefaultBtn'],
-                    parent: content,
-                });
-                const ok = Entry.Dom('div', {
-                    classes: ['popupOkBtn', 'popupDefaultBtn'],
-                    parent: content,
-                });
-                (text1 as any).text(Lang.Msgs.hw_download_text1);
-                (text2 as any).html(Lang.Msgs.hw_download_text2);
-                (text3 as any).text(Lang.Msgs.hw_download_text3);
-                (text4 as any).text(Lang.Msgs.hw_download_text4);
-                (cancel as any).text(Lang.Buttons.cancel);
-                (ok as any).html(Lang.Msgs.hw_download_btn);
-
-                (content as any).bindOnClick('.popupDefaultBtn', function() {
-                    const $this = $(this);
-                    if ($this.hasClass('popupOkBtn')) {
-                        hw.downloadConnector();
-                    }
-
-                    hw.popupHelper.hide('hwDownload');
-                });
-
-                popup.append(content);
-            },
-        });
     }
 
     private _convertHexToString(num: number | string) {
